requirements/requirements.txt
django==3.2.16
django-environ==0.9.0
mysqlclient==1.4.6
djangorestframework==3.14.0
django-cors-headers==3.14.0
debugpy==1.6.6
django-seed==0.3.1
psycopg2==2.9.5
pymysql==1.0.2
djangorestframework-simplejwt==5.2.2